pyproject.toml
[project]
name = "mcp-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.9.1",
    "openai>=1.82.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "networkx>=3.2.0",
    "plotly>=5.17.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "scipy>=1.16.0",
    "jieba>=0.42.1",
    "datasets>=4.0.0",
]
